pyproject.toml
[tool.poetry]
name = "djangoconus2022"
version = "0.1.0"
description = ""
authors = ["Simon Charette <charette.s@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
GitPython = "^3.1.27"
ipython = "^8.5.0"
black = "^22.8.0"
requests = "^2.28.1"
isort = "^5.10.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
